Update config.py
The info file generation is now generalized. Note that this is not
storing an info file on disk but is just used to serve unsharded
precomputed segmentation and raw volumes. SyConn backend is no longer
required to serve skeletons and meshes. Also, support for two datasets
of the j0126 acquisition are added.